fixup! misc: Add RP1 PIO driver
* Fixes build with Clang.

In file included from drivers/pwm/pwm-pio-rp1.c:20:
include/linux/pio_rp1.h:283:42: error: implicit conversion from 'uint' (aka 'unsigned int') to 'uint16_t' (aka 'unsigned short') changes value from 4294967295 to 65535 [-Werror,-Wconstant-conversion]
  283 |         args.origin = (program->origin == -1) ? PIO_ORIGIN_ANY : program->origin;
      |                     ~                           ^~~~~~~~~~~~~~
include/linux/pio_rp1.h:44:27: note: expanded from macro 'PIO_ORIGIN_ANY'
   44 | #define PIO_ORIGIN_ANY                  ((uint)(~0))
      |                                          ^~~~~~~~~~
include/linux/pio_rp1.h:312:42: error: implicit conversion from 'uint' (aka 'unsigned int') to 'uint16_t' (aka 'unsigned short') changes value from 4294967295 to 65535 [-Werror,-Wconstant-conversion]
  312 |         args.origin = (program->origin == -1) ? PIO_ORIGIN_ANY : program->origin;
      |                     ~                           ^~~~~~~~~~~~~~
include/linux/pio_rp1.h:44:27: note: expanded from macro 'PIO_ORIGIN_ANY'
   44 | #define PIO_ORIGIN_ANY                  ((uint)(~0))
      |                                          ^~~~~~~~~~
2 errors generated.

+++ include/linux/pio_rp1.h
@@ -41,7 +41,7 @@
 
 #define NUM_PIO_STATE_MACHINES		4
 #define PIO_INSTRUCTION_COUNT		32
-#define PIO_ORIGIN_ANY			((uint)(~0))
+#define PIO_ORIGIN_ANY			((uint16_t)(~0))
 #define GPIOS_MASK			((1 << RP1_PIO_GPIO_COUNT) - 1)
 
 #define PICO_NO_HARDWARE		0
